Fold `registryApi.ts` into `appChanges.ts`
Also move some functions from `fileUtil.ts` into `appChanges.ts` because
that is the only place where they are used and where they make more
sense.
